25 fetch endpoints from backend

+++ backend/src/resolverMap.ts
@@ -2,6 +2,9 @@ import { IResolvers } from 'graphql-tools';
 import { Post } from './models/post';
 import { User } from './models/user';
 import { Comment } from './models/comment';
+import bcrypt from 'bcryptjs';
+import { signToken } from './auth';
+import { AuthenticationError, UserInputError } from 'apollo-server-express';
 
 export const resolvers: IResolvers = {
   Query: {
@@ -31,9 +34,24 @@ export const resolvers: IResolvers = {
     },
     getComments: async (_, { postID }) => {
       try {
-        return await Comment.find({ parentId: postID });
+        return await Comment.find({ parentID: postID }).sort({ createdAt: -1 });
       } catch (err) {
         throw new Error('Error fetching comments');
+      }
+    },
+    getPostsByIds: async (_, { ids }) => {
+      try {
+        return await Post.find({ _id: { $in: ids } }).sort({ createdAt: -1 });
+      } catch (err) {
+        throw new Error('Error fetching posts by IDs');
+      }
+    },
+
+    getCommentsByIds: async (_, { ids }) => {
+      try {
+        return await Comment.find({ _id: { $in: ids } }).sort({ createdAt: -1 });
+      } catch (err) {
+        throw new Error('Error fetching comments by IDs');
       }
     },
     async searchAll(_: any, { query }: { query: string }) {
@@ -60,36 +78,91 @@ export const resolvers: IResolvers = {
   },
 
   Mutation: {
-    createPost: async (_, { body, author }) => {
+    createPost: async (_, { body }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to create a post');
+      }
+
+      const user = await User.findById(context.user.id);
+
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
+
       try {
-        const newPost = new Post({ body, author });
-        return await newPost.save();
+        const newPost = new Post({ body, author: user.username });
+        const savedComment = await newPost.save();
+
+        user.postIds.push(savedComment.id);
+        await user.save();
+
+        return savedComment;
       } catch (err) {
         throw new Error('Error creating post');
       }
     },
-    createUser: async (_, { username }) => {
-      try {
-        const existingUser = await User.findOne({ username });
-        if (existingUser) {
-          throw new Error('Username already taken');
-        }
-        const newUser = new User({ username });
-        return await newUser.save();
-      } catch (err) {
-        throw new Error(`Error creating user: ${(err as Error).message}`);
+    register: async (_, { username, password }) => {
+      const existingUser = await User.findOne({ username });
+      if (existingUser) {
+        throw new Error('Username already exists');
       }
+      const user = new User({ username, password });
+      await user.save();
+      const token = signToken(user);
+      return token;
+    },
+    login: async (_, { username, password }) => {
+      const user = await User.findOne({ username });
+      if (!user) {
+        throw new Error('Invalid credentials');
+      }
+      const valid = await user.comparePassword(password);
+      if (!valid) {
+        throw new Error('Invalid credentials');
+      }
+      const token = signToken(user);
+      return token;
     },
 
-    createComment: async (_, { body, author, parentID }) => {
+    createComment: async (_, { body, parentID }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to create a comment');
+      }
+
+      const user = await User.findById(context.user.id);
+
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
+
       try {
-        const newComment = new Comment({ body, author, parentID: parentID });
-        return await newComment.save();
+        const newComment = new Comment({ body, author: user.username, parentID: parentID });
+        const savedComment = await newComment.save();
+        await Post.findByIdAndUpdate(parentID, { $inc: { amtComments: 1 } });
+
+        user.commentIds.push(savedComment.id);
+        await user.save();
+
+        return savedComment;
       } catch (err) {
         throw new Error('Error creating comment');
       }
     },
-    deletePost: async (_, { id }) => {
+    deletePost: async (_, { id }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to create a comment');
+      }
+
+      const user = await User.findById(context.user.id);
+
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
+
+      if (!user.postIds.includes(id)) {
+        throw new AuthenticationError('You are not authorized to delete this post');
+      }
+
       try {
         const deletedPost = await Post.findByIdAndDelete(id);
         if (!deletedPost) {
@@ -102,71 +175,83 @@ export const resolvers: IResolvers = {
         throw new Error(`Error deleting post and its comments: ${(err as Error).message}`);
       }
     },
+    deleteComment: async (_, { id }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to delete a comment');
+      }
 
-    deleteComment: async (_, { id }) => {
+      const user = await User.findById(context.user.id);
+
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
+
+      if (!user.commentIds.includes(id)) {
+        throw new AuthenticationError('You are not authorized to delete this comment');
+      }
+
       try {
         const deletedComment = await Comment.findByIdAndDelete(id);
         if (!deletedComment) {
           throw new Error('Comment not found');
         }
+        await Post.findByIdAndUpdate(deletedComment.parentID, { $inc: { amtComments: -1 } });
+
         return deletedComment;
       } catch (err) {
         throw new Error(`Error deleting comment: ${(err as Error).message}`);
       }
     },
-    likePost: async (_, { postID }, { username }) => {
-      try {
-        const user = await User.findOne({ username });
-        if (!user) throw new Error('User not found');
 
-        if (user.likedPostIds.includes(postID)) {
-          throw new Error('Post already liked by this user');
-        }
+    likePost: async (_, { postID }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to like a post');
+      }
 
-        const post = await Post.findById(postID);
-        if (!post) throw new Error('Post not found');
+      const post = await Post.findById(postID);
+      if (!post) {
+        throw new UserInputError('Post not found');
+      }
 
+      const user = await User.findById(context.user.id);
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
+
+      if (!user.likedPostIds.includes(postID)) {
         post.amtLikes += 1;
         user.likedPostIds.push(postID);
-
         await post.save();
         await user.save();
-
-        return post;
-      } catch (err) {
-        if (err instanceof Error) {
-          throw new Error(`Error liking post: ${err.message}`);
-        } else {
-          throw new Error('Error liking post');
-        }
       }
+
+      return post;
     },
-    unlikePost: async (_, { postID }, { username }) => {
-      try {
-        const user = await User.findOne({ username });
-        if (!user) throw new Error('User not found');
 
-        if (!user.likedPostIds.includes(postID)) {
-          throw new Error('Post not liked by this user');
-        }
+    unlikePost: async (_, { postID }, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in to unlike a post');
+      }
 
-        const post = await Post.findById(postID);
-        if (!post) throw new Error('Post not found');
+      const post = await Post.findById(postID);
+      if (!post) {
+        throw new UserInputError('Post not found');
+      }
 
-        post.amtLikes -= 1;
-        user.likedPostIds.push(postID);
+      const user = await User.findById(context.user.id);
+      if (!user) {
+        throw new UserInputError('User not found');
+      }
 
+      const likedIndex = user.likedPostIds.indexOf(postID);
+      if (likedIndex > -1) {
+        post.amtLikes = Math.max(post.amtLikes - 1, 0);
+        user.likedPostIds.splice(likedIndex, 1);
         await post.save();
         await user.save();
-
-        return post;
-      } catch (err) {
-        if (err instanceof Error) {
-          throw new Error(`Error liking post: ${err.message}`);
-        } else {
-          throw new Error('Error liking post');
-        }
       }
+
+      return post;
     },
   },
   SearchResult: {
